Create README.md
This report explores the application of Convolutional Neural Networks (CNNs) in vegetable image classification, aiming to evaluate their efficacy using a dataset of 21,000 images across 15 vegetable classes. After preprocessing the data and configuring the model, training is conducted, resulting in an 81% accuracy on the test set. The model's potential applications in agriculture, such as quality control and inventory management, are highlighted, with suggestions for future improvements including advanced architectures and transfer learning techniques. Overall, this study underscores the CNN's effectiveness in accurately categorizing vegetable images, paving the way for broader deployment and refinement in agricultural applications.